Aenderung von log2 auf log
