+++ class/log.class.php
@@ -1,60 +1,62 @@
-<?php 
+<?php
 
 class log
 {
-	const LOG_PATH = __DIR__.'/../logs/';
-	
-	protected $entrys = array();
-	
-	protected function add($level, $desc, $array = NULL)
-	{
-		$this->entrys[] = array('date' => date("d.m.y H:i:s",time()), 'level' => $level, 'desc' => $desc, 'array' => $array);
-		return true;
-	}
-	
-	public function debug($desc, $array = NULL)
-	{
-		return $this->add(__FUNCTION__, $desc, $array);
-	}
-	
-	public function notice($desc, $array = NULL)
-	{
-		return $this->add(__FUNCTION__, $desc, $array);
-	}
-	
-	public function info($desc, $array = NULL)
-	{
-		return $this->add(__FUNCTION__, $desc, $array);
-	}
-	
-	public function warning($desc, $array = NULL)
-	{
-		return $this->add(__FUNCTION__, $desc, $array);
-	}
-	
-	public function error($desc, $array = NULL)
-	{
-		return $this->add(__FUNCTION__, $desc, $array);
-	}
-	
-	
-	public function __destruct()
-	{
-		$text = "";
-		foreach ($this->entrys as &$row)
-		{
-			$text .= $row["date"]."|| ".$row["level"]."|| ".$row["desc"]."|| ".print_r($row["array"],1)."\n";
-		}
-						
-		$timestamp = time();
-		$datum = date("Y-m-d",$timestamp);
-		$datumMonat = date("Y-m",$timestamp);
-		$file = self::LOG_PATH.$datumMonat."_logfolder/"."log-".$datum.".txt";
-		
-		if (!file_exists($file)) $fp = fopen($file,"x+");
-				
-		$fp = fopen($file,"a+");
-		fwrite($fp, $text);
-		fclose($fp);		
-	}
+    protected static function add($level, $desc, $context = NULL, $param = NULL)
+    {
+        $dir = __DIR__.'/../logs/';
+        $file = $dir.'syslog_'.date("ymd").'_1SRV.log';
+
+        $text = str_pad('['.date("r",time())."] [$level]", 41);
+                    $context = strpos($context, ')') === false ? $context.'()' : $context;
+            $text = str_pad("$text $context;", 120)."// $desc";
+
+            if (\is_array($param) && count($param) == 1)
+            {
+                foreach ($param as $name => $value)
+                {
+                    $text .= " | $name: $value";
+                }
+            }
+            elseif (is_array($param) || is_object($param))
+            {
+                //$text .= ' | '.json_encode($param);
+                //$text .= "\n".print_r($param, 1)."\n";
+            }
+            elseif (!empty($param))
+            {
+                $text .= " | $param";
+            }
+        $text .= "\n";
+
+        $fp = fopen($file, 'a');
+        fwrite($fp, $text);
+        fclose($fp);
+        return true;
+    }
+
+    public static function debug($desc, $context, $array = NULL)
+    {
+        return self::add(__FUNCTION__, $desc, $context, $array);
+    }
+
+    public static function notice($desc, $context, $array = NULL)
+    {
+        return self::add(__FUNCTION__, $desc, $context, $array);
+    }
+
+    public static function info($desc, $context, $array = NULL)
+    {
+        return self::add(__FUNCTION__, $desc, $context, $array);
+    }
+
+    public static function warning($desc, $context, $array = NULL)
+    {
+        return self::add(__FUNCTION__, $desc, $context, $array);
+    }
+
+    public static function error($desc, $context, $array = NULL)
+    {
+        return self::add(__FUNCTION__, $desc, $context, $array);
+    }
 }
